Update tính năng route giữa 2 điểm

+++ app/src/main/java/com/example/doan/FragmentMap.java
@@ -530,7 +530,9 @@ public class FragmentMap extends Fragment
                 LocationComponentPlugin locationComponentPlugin = getLocationComponent(mapView);
                 locationComponentPlugin.removeOnIndicatorPositionChangedListener(onIndicatorPositionChangedListener);
 
-                pointAnnotationManager.deleteAll();
+                if(FirstSearchOrigin==false){
+                    pointAnnotationManager.deleteAll();
+                }
 
                 // Tạo annotation mới tại vị trí click
                 PointAnnotationOptions pointAnnotationOptions = new PointAnnotationOptions()
